WebCrawler.py
The web crawler project is a script that crawls the internet and saves the content of specific URLs. Using multi-thread principles in its design, a web crawler is very useful for finding up-to-date information. A crawler bot is created with Python's request module or Scrapy, a Python open-source online crawling framework designed specifically for web scraping and data extraction via APIs. This python project's source code can be found here.